wip: get-campaigns-meta
